requirements.txt
langchain-core>=1.2.10,<2.0.0
langchain-community>=0.4.0,<0.5.0
langchain-classic>=1.0.0,<2.0.0
langchain-google-genai>=4.0.0,<5.0.0
langchain-text-splitters>=1.1.0,<2.0.0
pinecone>=5.1.0
python-dotenv>=1.0.0
pypdf>=4.0.0
